Add tools tests and documentation.
