+++ Sources/PackageDescription/TapestryConfig.swift
@@ -2,7 +2,9 @@ import Foundation
 
 // MARK: - TapestryConfig
 
+/// Describes Tapestry configuration
 public struct TapestryConfig: Codable {
+    /// Define your release steps
     public let release: Release?
 
     public init(release: Release? = nil) {
