+++ Sources/TapestryCore/Constants.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+public class Constants {
+    public static let gitRepositoryURL = "https://github.com/AckeeCZ/tapestry.git"
+    public static let version = "0.0.2"
+    public static let swiftVersion: String = "5.1"
+}
