+++ Sources/TapestryCore/XcodeController.swift
@@ -1,10 +1,15 @@
 import Basic
 import class TuistCore.System
 
+/// Interact with xcode CLI tools
 public protocol XcodeControlling {
+    /// Opens file with xcode
+    /// - Parameters:
+    ///     - path: Describes path of the file to open
     func open(at path: AbsolutePath) throws
 }
 
+/// Interact with xcode CLI tools
 public final class XcodeController: XcodeControlling {
     /// Shared instance
     public static var shared: XcodeControlling = XcodeController()
