+++ Sources/TapestryGen/TapestryConfig.swift
@@ -1,6 +1,8 @@
 import Basic
 
+/// Describes Tapestry configuration
 public struct TapestryConfig {
+    /// Define your release steps
     public let release: Release
     
     public init(release: Release) {
